+++ src/devices/src/virtio/net/device.rs
@@ -229,6 +229,7 @@ impl Net {
         // If limiter.consume() fails it means there is no more TokenType::Ops
         // budget and rate limiting is in effect.
         if !self.rx_rate_limiter.consume(1, TokenType::Ops) {
+            METRICS.net.rx_rate_limiter_throttled.inc();
             return false;
         }
         // If limiter.consume() fails it means there is no more TokenType::Bytes
@@ -239,6 +240,7 @@ impl Net {
         {
             // revert the OPS consume()
             self.rx_rate_limiter.manual_replenish(1, TokenType::Ops);
+            METRICS.net.rx_rate_limiter_throttled.inc();
             return false;
         }
 
@@ -488,6 +490,7 @@ impl Net {
                 // Stop processing the queue and return this descriptor chain to the
                 // avail ring, for later processing.
                 tx_queue.undo_pop();
+                METRICS.net.tx_rate_limiter_throttled.inc();
                 break;
             }
 
@@ -516,6 +519,7 @@ impl Net {
                 // Stop processing the queue and return this descriptor chain to the
                 // avail ring, for later processing.
                 tx_queue.undo_pop();
+                METRICS.net.tx_rate_limiter_throttled.inc();
                 break;
             }
 
@@ -605,6 +609,8 @@ impl Net {
             // If the limiter is not blocked, resume the receiving of bytes.
             if !self.rx_rate_limiter.is_blocked() {
                 self.resume_rx().unwrap_or_else(report_net_event_fail);
+            } else {
+                METRICS.net.rx_rate_limiter_throttled.inc();
             }
         }
     }
@@ -623,6 +629,7 @@ impl Net {
 
         // While limiter is blocked, don't process any more incoming.
         if self.rx_rate_limiter.is_blocked() {
+            METRICS.net.rx_rate_limiter_throttled.inc();
             return;
         }
 
@@ -646,6 +653,8 @@ impl Net {
         // If the limiter is not blocked, continue transmitting bytes.
         {
             self.process_tx().unwrap_or_else(report_net_event_fail);
+        } else {
+            METRICS.net.tx_rate_limiter_throttled.inc();
         }
     }
 
@@ -1428,6 +1437,7 @@ pub(crate) mod tests {
         net.rx_rate_limiter = RateLimiter::new(0, 0, 0, 0, 0, 0).unwrap();
         let rate_limiter_event =
             EpollEvent::new(EventSet::IN, net.rx_rate_limiter.as_raw_fd() as u64);
+        // There is no actual event on the rate limiter's timerfd.
         check_metric_after_block!(
             &METRICS.net.event_fails,
             1,
@@ -1448,6 +1458,7 @@ pub(crate) mod tests {
         let rate_limiter_event =
             EpollEvent::new(EventSet::IN, net.tx_rate_limiter.as_raw_fd() as u64);
         net.process(&rate_limiter_event, &mut event_manager);
+        // There is no actual event on the rate limiter's timerfd.
         check_metric_after_block!(
             &METRICS.net.event_fails,
             1,
@@ -1492,6 +1503,7 @@ pub(crate) mod tests {
 
                 // assert that limiter is blocked
                 assert!(net.tx_rate_limiter.is_blocked());
+                assert_eq!(METRICS.net.tx_rate_limiter_throttled.count(), 1);
                 // make sure the data is still queued for processing
                 assert_eq!(txq.used.idx.get(), 0);
             }
@@ -1502,9 +1514,9 @@ pub(crate) mod tests {
 
             // following TX procedure should succeed because bandwidth should now be available
             {
-                // tx_count increments 1 from process_tx() and 1 from write_to_mmds_or_tap()
                 let tx_limiter_event =
                     EpollEvent::new(EventSet::IN, net.tx_rate_limiter.as_raw_fd() as u64);
+                // tx_count increments 1 from process_tx() and 1 from write_to_mmds_or_tap()
                 check_metric_after_block!(
                     &METRICS.net.tx_count,
                     2,
@@ -1543,6 +1555,7 @@ pub(crate) mod tests {
 
                 // assert that limiter is blocked
                 assert!(net.rx_rate_limiter.is_blocked());
+                assert_eq!(METRICS.net.rx_rate_limiter_throttled.count(), 1);
                 assert!(net.rx_deferred_frame);
                 // assert that no operation actually completed (limiter blocked it)
                 assert_eq!(net.interrupt_evt.read().unwrap(), 2);
@@ -1560,7 +1573,12 @@ pub(crate) mod tests {
                 net.interrupt_evt.write(1).unwrap();
                 let rx_limiter_event =
                     EpollEvent::new(EventSet::IN, net.rx_rate_limiter.as_raw_fd() as u64);
-                net.process(&rx_limiter_event, &mut event_manager);
+                // no longer throttled
+                check_metric_after_block!(
+                    &METRICS.net.rx_rate_limiter_throttled,
+                    0,
+                    net.process(&rx_limiter_event, &mut event_manager)
+                );
                 // validate the rate_limiter is no longer blocked
                 assert!(!net.rx_rate_limiter.is_blocked());
                 // make sure the virtio queue operation completed this time
@@ -1607,7 +1625,11 @@ pub(crate) mod tests {
                 net.queue_evts[TX_INDEX].write(1).unwrap();
                 let tx_event =
                     EpollEvent::new(EventSet::IN, net.queue_evts[TX_INDEX].as_raw_fd() as u64);
-                net.process(&tx_event, &mut event_manager);
+                check_metric_after_block!(
+                    METRICS.net.tx_rate_limiter_throttled,
+                    1,
+                    net.process(&tx_event, &mut event_manager)
+                );
 
                 // assert that limiter is blocked
                 assert!(net.tx_rate_limiter.is_blocked());
@@ -1623,7 +1645,12 @@ pub(crate) mod tests {
             {
                 let tx_rate_limiter_event =
                     EpollEvent::new(EventSet::IN, net.tx_rate_limiter.as_raw_fd() as u64);
-                net.process(&tx_rate_limiter_event, &mut event_manager);
+                // no longer throttled
+                check_metric_after_block!(
+                    &METRICS.net.tx_rate_limiter_throttled,
+                    0,
+                    net.process(&tx_rate_limiter_event, &mut event_manager)
+                );
                 // validate the rate_limiter is no longer blocked
                 assert!(!net.tx_rate_limiter.is_blocked());
                 // make sure the data queue advanced
@@ -1653,10 +1680,16 @@ pub(crate) mod tests {
                 net.interrupt_evt.write(1).unwrap();
                 // trigger the RX handler
                 let rx_event = EpollEvent::new(EventSet::IN, net.tap.as_raw_fd() as u64);
-                net.process(&rx_event, &mut event_manager);
+                check_metric_after_block!(
+                    METRICS.net.rx_rate_limiter_throttled,
+                    1,
+                    net.process(&rx_event, &mut event_manager)
+                );
+                //net.process(&rx_event, &mut event_manager);
 
                 // assert that limiter is blocked
                 assert!(net.rx_rate_limiter.is_blocked());
+                assert!(METRICS.net.rx_rate_limiter_throttled.count() >= 1);
                 assert!(net.rx_deferred_frame);
                 // assert that no operation actually completed (limiter blocked it)
                 assert_eq!(net.interrupt_evt.read().unwrap(), 2);